Crate lyon_algorithms

2d Path transformation and manipulation algorithms.

This crate is reexported in lyon.

Re-exports

pub extern crate lyon_path as path;

pub use crate::path::geom;

Modules

aabb

Bounding rectangle computation for paths.

area

Approximate the area of a path.

fit

Fit paths into rectangles.

hatching

Generate hatching and dotted patterns in a path.

hit_test

Determine whether a point is inside a path.

length

Approximate path length.

math

f32 version of the lyon_geom types used everywhere. Most other lyon crates reexport them.

measure

Perform cached measurements and split operations on a path.

raycast

Find the first collision between a ray and a path.

rect

rounded_polygon

walk

Move at a defined speed along a path.

winding